Separate vessel colors assignment from backend request
