+++ api-server/ui/backend-requests.ts
@@ -1,30 +1,15 @@
 import { fromPromise, Observable } from 'baconjs'
-import Color = require('color')
 import Debug from 'debug'
-import palette from 'google-palette'
 
 import { toQueryString } from '../domain/Geo'
 import { VesselData, VesselId } from '../domain/Vessel'
-import { LoadedTrack, Vessel, Viewport } from './ui-domain'
+import { LoadedTrack, Viewport } from './ui-domain'
 
 const debug = Debug('stash:backend-requests')
 
-export function loadVessels(): Observable<Vessel[]> {
+export function loadVessels(): Observable<VesselData[]> {
   debug('Loading vessels..')
-  return fromPromise(
-    fetch(`/contexts`)
-      .then(res => res.json())
-      .then((vessels: VesselData[]) => {
-        const colors = palette('mpn65', vessels.length)
-        return vessels.map((v, idx) => ({
-          vesselId: v.vesselId,
-          name: v.name,
-          trackColor: Color(`#${colors[idx % colors.length]}`)
-            .desaturate(0.5)
-            .lighten(0.06)
-        }))
-      })
-  )
+  return fromPromise(fetch(`/contexts`).then(res => res.json()))
 }
 
 async function loadTrack(
